Add TS textobjects and consolidat TS config
An attempt to extend function.outer to incluse comments had been made,
but it's ultimately fails due to weird quantifier handling. Quantifiers
don't fail per se, but '*' one is non-greedy, making it impossible to
match and capture the first line of multiline comments.

An example query is commented out and waits for potential fix in neovim.