Move solargraph setup outside of mason
Mason install will still work, but I can't keep relying on it.
While move from lsp-installer to mason fixed GEM_HOME issues,
the resulting installation doesn't work well with version managers.

In other words, I gave up after having to reinstall solargraph thrice
when it insisted to run on a ruby version that didn't match the one
I was using in the project. This is not Mason fault, but rather version
management one.