Update git conflict keybindings
I couldn't find a way to make `vim.keymap.del` to silently do nothing if
keymap in question doesn't exist. I'm not interested to know that the keymap
didn't exist before the call, but rather that it definitely doesn't exist
after.

While not a big deal in a sequential execution, in an evented system I
simply don't know whether the call to set up the keymaps was ever
issued. Turns out, git-conflict can detect that the conflict is ongoing,
but the resolution selection has already taken place. In which case it
doesn't issue the event that leads to us setting up the keymaps.

So in order to not see the errors in such a case, I'm looping over
buffer keymaps and removing the ones that match git conflict right hand
side.