Cleanup TS and bodge together :Inspect in a popup
I'm gradually preparing for neovim 0.10. As a part of that, I wanted to
get :Inspect in a float, which led me to copying a hell of a lot of code
from :Inspect implementation to populate the floating window of my
liking.

I can probably do it better now that I've done it once, will try again
later.